Change blacklight gem versions from using :git option to pull from github directly, to using (compatible) semantic versions. NOTE: this requires pinning kaminari to < 0.14.0. Fix this later after upgrading blacklight.
Change jetty ports to 8983 (which is same as default). I think the 8181 ports are result of a dev setting making it into production.

I'd like to use the same port everywhere unless there is some issue with that. And might as well make it the default. Trying it here, to be tested on staging.