* Removes a bunch more solr data from revision. * Changes config/solr.yml URL for development env to point back develoopment core on localhost:8983. * Changes config/jetty.yml to specify port 8181 for production env when starting jetty w/ rake jetty:start.
NOTE: There is a special script for starting jetty on production (i.e. sudo /sbin/service jetty-bostonlocaltv).
However, this is a big long bash script that is currently not in revision, and I'm wondering why we decided to
go with that instead of using the rake tasks provided by jettywrapper. So changing config/jetty.yml for
production environment may be a moot point, but if it works, I would prefer to use that method of starting/stopping
jetty rather than using /sbin/service.