Admin can make users sponsors from admin interface
Within the Artifact view in activeadmin interface, an admin can "confirm
sponsorship", essentially indicating that a user has paid for the
digitization of an artifact. The sponsorship confirmation can also be
undone/uncomfirmed. Add tests for the Artifact admin interface.
Configure database_cleaner to actually clear out the db for tests.