Sponsorships removed from user on request withdraw
Uncovered this weird bug that blocked me for part of the day: upon
request_withdraw on an artifact, the sponsorship/user was removed from
the artifact itself, but it wasn't being removed from the user, which
was causing an issue at the view layer (on dashboard) in certain edge
cases. Now we are reloading the relationship immediately upon withdraw
which resolves the issue I was experiencing.