Refactored and fixed artifact/catalog/show page
There was a lot of logic wrapped up in which links and messages to
display on an artifact page and tonight I discovered it wasn't working
as intended 100% of the time. I refactored all that logic out into a
helper. Should make things easier down the road too, if we ever need to
change this up.