WIP tests for artifact show logic
These are not working at the moment - spent maybe too much time trying
to figure out how to stub out the blacklight stuff so I could get to a
point in the test where I could actually visit the catalog show path in
order to test all the work I did in the previous commit, regarding the
refactor of the logic related to the artifact/catalog show feature. I
will need to connect with Drew on this tomorrow if we do want to move
forward with tests for this, because I wasn't able to get it quite where
I needed to.