Fixes #93 admins can change artifact to published
In admin interface, an admin can 'publish' an artifact. This changes the
state of the artifact to published, reflects the updated status in the
admin interface as well as the user's dashboard, and emails any
associated users to notify them of the artifact's availability. Tests
for this are coming next.